+++ test/currency_textfield_test.dart
@@ -76,4 +76,11 @@ void main() {
     final controller = CurrencyTextFieldController(initIntValue: 195,currencyOnLeft: false);
     expect(controller.text, '1,95 R\$');
   });
+
+  test('negative_and_block', () {
+    final controller = CurrencyTextFieldController(initIntValue: -195);
+    final controller2 = CurrencyTextFieldController(initIntValue: -195,enableNegative: false);
+    expect(controller.text, '-R\$ 1,95');
+    expect(controller2.text, 'R\$ 1,95');
+  });
 }