+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [2.9.0]
+- Added the possibility to have negative numbers
+- New `enableNegative` parameter to block user from setting negative numbers
+
 ## [2.8.0]
 - Added currencyOnLeft parameter: now you can decide if the symbol will be before or after the number
 - breaking change: because of the new parameter, `leftSymbol` was renamed to `currencySymbol`